Add clone impl to filters (#13)
This is handy in some cases, for things like partitioned reads.

In general, the cost of adding a clone impl like this is a bit of
compilation time (very little), and of course the runtime cost if you
actually use it. For filters, however, the cost of using them is
generally negligible compared to the cost of the request itself.