Custom auth headers (#15)
This is useful to allow users to set auth headers in alternative ways. This could have been done with a trait, but that would have introduced a performance cost for each call, even when using just ApiKey or OIDCToken.

Using an enum and wrapping the inevitable dynamic dispatch inside a variant makes the cost much smaller, and is still as flexible.