Alternative file upload (#21)
When adding file upload we ran into the problem that azure does not support content-encoding: chunked, which is the default when uploading a stream.

The issue is that when uploading a stream, the total size is not known, so we have to chunk the upload. If, however, we specify the content-length before upload, we are able to stream data without using chunked encoding. This is presumably how python does it.

This does of course require knowing the size of what you're uploading before hand. As we don't include tokio as dependency for the SDK, this means just adding a new method upload_stream_known_size, which takes the size of the stream in bytes.