Unify into a single Items type (#216)
There are multiple different types (`Items`, `ItemsWithCursor`, `ItemsWithoutCursor`, `ItemsWithIgnoreUnknownIds`) that all represent the same basic structure: `{"items":[...], ...}`. The reduce code duplication, I'm proposing that this is changed to have a single base type (`Items`) which is generic enough to work with all of these cases:

```
#[derive(Copy, Clone, Debug, Default, Serialize, Deserialize)]
pub struct Items<C, E = ()> {
    pub items: C,
    #[serde(flatten)]
    pub extra_fields: E,
}
```

The `items` field should be a collection of items, such as `Vec<T>` or `&[T]`.

The `extra_fields` field can be used to extend the list object with additional fields. In this PR, two additional types are defined for this purpose: `Cursor`, which adds the `nextCursor` field, and `IgnoreUnknownIds`, which adds the `ignoreUnknownIds` field.

In addition, `ItemsVec<T, E = ()>` is added as a convenience alias for `Items<Vec<T>, E>`, effectively replacing the existing `ItemsWithoutCursor`.