Make the crate build for the wasm32-unknown-unknown target
Requires an unreleased version of reqwest-middleware, as well as a great deal of
hackery to avoid send bounds since JS futures are not send.

Reqwest on wasm also doesn't support _sending_ streams, for whatever
reason, so that feature is disabled, we load the entire file to be sent
into memory, with all that entails. This may be possible to fix in the
future.