Make error in authenticator a little clearer (#282)
There isn't a ton we can do to make it obvious what went wrong, since we can't really log the response from an OAuth request, but we can at least make it obvious that the error came from the authenticator. Currently you'll just get "Error in middleware: Failed to deserialize" which is less helpful.