Support for data modeling records (#288)
* Support for data modeling records

Not the aggregate API yet, there's a separate PR for just the types for
that, but the other three APIs are defined here.

Testing these are hard, since they are eventually consistent, and you
need a limited time window when filtering (I want my tests to be
reproducible, even on a new CDF project. I dislike hard-coding known
timestamps for a specific project).

* Fix after merge

* Update cognite/src/api/data_modeling/records.rs

Co-authored-by: Ikechukwu Eze <92305839+ikezedev@users.noreply.github.com>

---------

Co-authored-by: Ikechukwu Eze <92305839+ikezedev@users.noreply.github.com>