Utility to stream datapoints from CDF
We have something like this in the python SDK, and we have similar
mechanisms for partitioned reads for certain other resources.

It is useful to be able to retrieve _all_ datapoints for a set of
timeseries, and not worry about batching, parallelism, or following
cursors. This implements a stream-based wrapper for doing these queries,
with two public variants: One for getting batches of datapoints as they
are retrieved, and a wrapper around that which streams individual
datapoints.