Wrap the authenticator in an arc (#313)
* Wrap the authenticator in an arc

This is slightly breaking, though I think most users will just use
methods on the builder, so it shouldn't affect most people.

I need this for a project where I use the same OAuth token for REST and
something else, which is admittedly niche, but treating the
authenticator as something "shared" is correct in that it is only ever
used that way.

* Make callbacks arc as well