Add utility for list of identities (#307)
When working with core APIs, it's often common to just have a list of
external IDs, or a list of internal IDs. Right now, you have to convert
that into a list of Identity or IdentityOrInstance, which is a little
annoying, and not actually necessary just for serialization.

This is a bit of an experiment to help with this, by adding an
IdentityList type that wraps a list of strings, integers, etc. and
serializes them as CDF IDs. For now just adding it to APIs that use a
few core retrieve traits.